Fix bug reported by Stephan Merz:
"For the first spec that I loaded, I got an error because I had
forgotten to include the library repository for the prover library. So I
added that folder to the search path in the TLA+ Preferences, and when I
clicked "Apply", the Toolbox told me that the spec had to be reparsed,
to which I agreed. However, the window displaying the parse error did
not disappear (I presume the spec was indeed reparsed but the windows
weren't updated). Manually reparsing the spec solved the problem.
Curiously, when I then removed the folder from the search path and
agreed to reparsing, the error was displayed immediately."

Technically the fix simplifies what appears to be legacy code with a
function provided by the ParserHelper. ParserHelper triggers sightly
different events in the Eclipse foundation's AutoBuilder code.