* Replaced org.lamport.tla.toolbox.util.StringHelper#trimFront with an inverted implementation similar to trimEnd. The new implementation does not rely on java.lang.String#split whose behavior changed in Java8.
* StringHelperTest is a collection of quick and dirty (read "far from
complete") test cases for StringHelper. It is intended to show a bug on
Java8 related to whitespace matching. It also reveals corner cases (on
Java7 and Java) where the StringHelper documentation unclear or where
the result is questionable in my opinion. These tests are commented for
the moment pending further investigation.