Add boolean JVM system property "tlc2.tool.liveness.statistics" to turn on liveness statistic gathering on demand.
On the command line pass "-Dtlc2.tool.liveness.statistics=true" to the
JVM, in the Toolbox copy the same string into JVM arguments on the
"Advanced Options" page of the Model editor.

TLC prints the final statistics at the end of the model checker run in
the Toolbox's user output area on the "Result Page" of the Model editor.

If statistics remain off, the code is clever enough to not collect
statistics at all but to use a DummyBucketStatistics instance which
ignores all invocations.