CloneCommandContributionItem has been added to replace a Eclipse foundation "enabledWhen" core expression to programmatically handle the enablement and visibility of CommandContributionItems.
The problem with the core expression is, that a change in behavior
introduced with 4.x causes the "Clone Model" submenu entries to
disappear
from the main menu. The reason is, that open and closing the Toolbox
spec
explorer context menu changes the "activeMenu" evaluated by the
expression framework from [] to "toolbox.explorer.popup" to nothing
(IEvaluationContext.UNDEFINED_VARIABLE). Once it is nothing, the
WithExpression determining if the CommandContributionItem is visible
always evaluates to false no matter what.

Because the CommandContributionItems are only rendered by the
"Clone Model" main menu item, which itself has an enabledWhen clause, we
don't really need the enabledWhen on the item level anyway.

Also see handler definition in
org.lamport.tla.toolbox.tool.tlc.ui/plugin.xml