Support updating (outgoing) transitions of an existing GraphNode. This requires DiskGraph#getNode(fp, tixd) to return an existing GraphNode iff a GraphNode with an identical fp and tixd pair has already been added to the graph.
Updating a GraphNode is used *in simulation mode only*. Suppose the
trace is fps <<1, 2, 1, 3, 5>> (with tidx being 1 for simplicity
reasons). The node with fp 1 first gets added to the DiskGraph with one
successor, namely 2. Then it's added again with successor 3. The old
implementation would replace successor 2 with 3 whereas this commit
changes DiskGraph to preserve successor 2 and add 3.
Technically the backing file of nodePtrTbl will contain two records for
GraphNode with fp 1. The first with a single successor 2, the second
with successors 2 and 3. When the file gets read to recover the
DiskGraph, it's read front to end replacing. It thus replaces the first
entry with the single successor with the correct one with two.