Make sure a GraphNode instance does *not* get added to the DiskGraph when the set of successors doesn't change.
This introduces mocking with EasyMock to tlatools. Evaluated Mockito too
which failed to create a mock of OrderOfSolutions and Tool. The reason
why the TLAToolbox.target referes to EasyMock 2.4, yet 3.3.1 has been
committed to the repository, is due to 3.3.1 missing from Orbit. This
does not matter much though. The code compiles with 2.4 fine, it just
needs 3.3.1 at runtime to create mocks out of classes (2.4 only seems to
support creating mocks from interfaces).

(Unfinalized methods which according to documentation, has no
performance drawbacks.