[Bugfix] Remove deadlock when finishAll & suspendAll race.
suspendAll from main is on the heels of finishAll, but not quite as
fast. SuspendAll's synchronized(this.mu) is executed one clock tick
after  finishAll's thus waiting for finishAll() to notify all waiters of
this.mu. With main being the only potential waiter in the system nobody
gets notified and the lock on this released subsequently. Now it's
suspendAll's turn. It acquires the lock on this.mu and immediately after
goes to wait on it (this.mu) to be waken by worker threads eventually.
Unfortunately, there are none left in the system. All have long
finishedAll.

The fix is to check the this.finish variable one more time directly
after suspendAll acquires the lock this.mu, it checks if it still has to
wait for workers. To make sure it reads the most recent value, it's
declared volatile to establish a happens-before relation (since Java5's
memory model) between workers and main. Otherwise the VM might decide to
let main read a cached value of false of this.finish.

The deadlock was found by NASA's Java Pathfinder (jpf) model checker. The
fix was then verified with jpf (up to the point where it ran out of mem).