[Feature] Add TLCIterator#getLast(long) where long acts as a lower bound for the sequential iteration through the (potentially large) in-memory buffer.
Starting with the second disk flush, the maximum value (fingerprint) of
the in-memory and on-disk file has to be calculated. The maximum on-disk
value is kept as a variable, while the in-memory value requires a lookup
by iterating the buffer end to front. The moment the iterator sees a
value smaller than the on-disk max value, it can stop its scan. It won't
find a value larger than the on-disk one.