[Tests] Add tests that exemplify where TLC fails to produce the shortest error trace for liveness checking.
System LOOP without fairness can stutter after the initial state, but
TLC produces a counter-example showing stuttering with two additional
states. This is due to LiveWorker#checkSccs(..) generating the SCC
before it checks the states on the SCC end to start.

Forcing TLC to check liveness after insertion of each state into the
behavior graph causes it to return the shortest counter-example.