Toolbox launch fails during tests.
The Toolbox initialization can be described as hackery with layers of workaround on top. This has finally proven unmaintainable which caused this larger refactoring.

It now checks the instance location at the correct place instead of in the Activator. Formerly it was done in the Activator, because it would run before the Application which was a result of an upside-down extension point. With the extension point reversed, the Application is executed prior to the Activator.

[Bug](Toolbox]