TLC unnecessarily permutes all TLCState's variables.
With state symmetry defined, TLC loops over all permutations and applies each permutation to all of the state's variables during the fingerprint calculation. However, it unnecessarily keeps applying the current permutation to the state's variables even if the first application shows that the resulting state is not closer to the smallest one. Thus, TLC can skip remaining variables preemptively once it is determined that the resulting state won't be closer to the smallest state under symmetry.

Experiments show that this enhancement significantly reduces the costs (three to eight times faster) to find the smallest state under symmetry when the complexity of TLCState is large (nested function/record/... variables). It has negligible runtime effects when the variables are simple.

[Performance][TLC][Changelog]