Functionality to re-disable the "Always run in background" preference.
The first time the Toolbox is run, the 'Always run in background'
preference is set to false and the preferences written to disk; at the
same time, we make our own preference setting letting our future selves
know that we have already twiddled this once. If the user chooses to
re-enable 'Always run in background' after this point, we will not
affect that setting.

Fixes issue #101