Performance regression due to bugfix for Bugzilla #279.
Calling SetEnumValue#convert causes e.g. an instance of type SubsetValue
to be explicitly enumerated. Prior to the fix for #279, TLC was smart
enough to not enumerate all elements a-priori but to generate elements
one-by-one and return the first for which the given predicate is true.

See commit 1fd974d5d2de462bf517c03455839bc082b0cf2f and
4a62e5b48c9c42265cf1836de901523b4399a733 on which this commit depends.

[Bug][TLC]