Configuration Parser: remove manual written ReInit() method
When tla2sany parses more than one spec all involved objects must
be in a fresh state for each spec. This is usally done by
instantiating a new object of a class or by calling ReInit() methods
to bring them in a fresh state.

Since the Configuration parser is static it seems that
a call to ReInit() must be done.
This is not true.
The static load() method, which is called for each spec,
instantiates a Configuration object each time it parses one
Configuration unit. Then this object is destroyed.

So this handwritten ReInit() method is not needed and can be removed.
Also it relys on structures of the generated code by JavaCC
which may change in future versions.

[Refactor][SANY]