Configuration Parser: remove a bugfix
It seems a bit odd to remove a bugfix, but if this bufix fixes a real bug
then it should be fixed in JavaCC and not in the generated code.
I was not able to trigger the bug and removing it made no difference
on my systems.

In the very unlikely case that this change breaks something,
it can easily be spotted with git bisect.

[Refactor][SANY]