Configuration Parser: use the static parser the right way
Remove a missuse of the static configuration parser.
Before the parser is called the source of the input must
be specified and passed to the ReInit() method to
(re-)initialize the internal members of the parser.
Thats the way how the JavaCC devs want us to use the parser
in the static version.

This patch removes the handwritten ReInit() method
which was used to do the reinitialization without
knowledge of the input and the two resulting
bugfixes which made that ReInit() method work,
but broke the way it should be used.

The code is now restored the way it is generated by JavaCC.

[Refactor][SANY]