Configuration Parser: input file: update imports and package name
Rename the package names from tlatk to tla2sany and
import the packages like they are imported in the file
tlatools/src/tla2sany/Configuration.java.

Note: When generating a parser from config.jj these imports are
copied to the two files Configuration.java and ConfigurationTokenManager.java.
The package name will be copied to all 7 files.

This changes are non-functional.

[Refactor][SANY]