Initial implementation of the trace animation feature
This patch adds initial support for animating TLC traces by providing a 'view' expression that produces a visual representation of a TLA+ state. The trace animator re-uses most of the same functionality from the trace explorer, except that it evaluates the 'view' expression at each state of a trace instead of user defined expressions. Each frame of an animation is an SVG element. These are then stitched together into an interactive animation and displayed in an HTML viewer in the Toolbox.