Annotation-based loading of TLC module overrides for TLA+ operators.
Limits
------

Contrary to TLA+ operators, a module override has no access to the
constants (CONSTANTS) and variables (VARIABLES) declared in the spec
unless given as parameters.  It is completely stateless.  In other
words, the scope of the module override is limited to the method's
parameters.

It is up to the Java implementation to correctly handle all possible
types of the input parameters.  Care must be taken in case an operator
accepts a higher-level operator.

Howto
-----

Annotate any public static method in any class with @TLAPlusOverride.
Annotation takes two arguments: Module name and operator identifier.
The number of method parameters is expect to match that of the TLA+
operator.  The types of the method parameters and the method's return
value is tlc2.value.IValue (or a sub-type).

Finally, create (one) index class tlc2.overrides.TLCOverrides that
implements the tlc2.overrides.ITLCOverrides interface. TLCOverrides#get
returns an array of all classes with overrides. This mechanism only
loads the first class definition named tlc2.overrides.TLCOverrides on
the Java classpath.

Look out for TLC startup warnings if the setup is incorrect.

Improvement
-----------

Compared to the old mechanism, overrides no longer need to be:
- placed in the default namespace
- class names aligned with module names
- method names aligned with TLA+ operators

Addresses GitHub issue #326
https://github.com/tlaplus/tlaplus/issues/326

[Feature][TLC][Changelog]