Declare and define _TETrace and _TEPosition trace expression operators in (new) Toolbox module.
The _TETrace and _TEPosition used to be generated as part of TE.tla
(trace expressions).  This made it impossible to form expressions based
on either one the two operators outside the Toolbox Trace Explorer. E.g.
it wasn't possible to define trace expressions built from _TETrace
and/or _TEPosition in the actual spec.  This however was desirable with
complex trace expressions for which the Trace Explorer UI is not ideal.

With this commit, the _TETrace operator defined in module Toolbox gets
overridden by a definition override defined in TE.tla during trace
exploration.  Similar to the previous revision, the
TraceExpressionModelWriter generates the definition override.

For existing trace expression - built from _TETrace
and/or _TEPosition - nothing changes.

Part of the work for GitHub issue #342
https://github.com/tlaplus/tlaplus/issues/342

[Feature][Toolbox][Changelog]