Fix TLC rerun at same JVM for some temporal property
I'm in a clojure REPL (but it's only a technicality). When running
the TLC model checker for some TLA file for the first time, it shows
me the correct error for a temporal property. But for a rerun of the
TLC model checker, it does not show me this error.

```clojure
;; Code which is run twice
(doto (tlc2.TLC.)
  (.handleParameters (into-array
                      ["/Users/feodrippe/dev/my-practical-tlaplus/ch1/wire.tla"]))
  (.process))
```

I was debugging it and found that there are some assumptions for the
initial values of the `LiveWorker` class. Maybe it should reset all the
static variables to their initial values when the method `process` is run,
but for now I've just created a public function to reset `errFoundByThread`
whenever someone needs it, I don't want to break anyone's code.

To give a more accurate time, you can also reset startTime at TLC class.