Errors during liveness checking do not terminate TLC.
The error/exception - such as an OutOfMemoryError - shows up on the
console, but does not terminate TLC.  A user thus might miss the
error/exception and incorrectly assume that the spec satisfies the
liveness properties.

The effect of this fix can be studied with unit test
CodePlexBug09EWD749FL2Test that terminates with a meaningful "ran
out of memory" message instead of showing TLC regular success
message despite going out of memory (run test with -Xmx56m).

Fixes Github issue #1
https://github.com/tlaplus/tlaplus/issues/1

[Bug][TLC][Changelog]