TLC shows no error trace on EvalExceptions thrown by module overrides.
Expressions that cause EvalExceptions in module overrides are silly
expressions like `x < TRUE`, but also a violated TLC!Assert, bogus
TLCGet/TLCSet, ....

TLC module overrides used to wrap nested EvalExceptions (the root cause)
in a (more technical) TLCRuntimeException, which was changed in commit
6569d03ca1be31ae1cf3cef0a23b005cbc6ecc3 to not wrap EvalExceptions. The
goal was to tell users the root cause. This, however, introduced a
regression because the global exception handling was not updated to
print an error-trace on EvalException.  Instead, the error on stdout
omitted the trace and only showed the stack trace pointing to the spec
location of the TLA+ expression.

Fixes Github issue #461
https://github.com/tlaplus/tlaplus/issues/461

Regression introduced with commit
66569d03ca1be31ae1cf3cef0a23b005cbc6ecc3
https://github.com/tlaplus/tlaplus/commit/66569d03ca1be31ae1cf3cef0a23b005cbc6ecc3

[Regression][Bug][TLC][Changelog]