Visualize the coverage between a spec's TLA actions when running simulation mode.
Let A be the set of TLA+ actions (technically TLC sub-actions) of a spec
and let S be the set of TLA+ states of a model.  Let R be the relation
(A \X S \X A). Let T be a multiset with domain R and range Int that
marks for each triple in R the number of times it was generated by TLC's
simulation mode (a trace <<a1, s1, a2, s2, a3, ...>> corresponds to
{<<a1,s1,a2>>,<<a2,s2,a3>>, ...}).  We construct a weighted, directed
graph G with G.nodes = A and G.edges = R and weights given by T[edge].

G -in dot notation- is periodically written to disk as part of progress
reporting.  Activate it with "-Dtlc2.tool.Simulator.actionStats=true".
The weights are visualized by (logarithmically) increasing the pen-width
of edges.  Dashed, green edges indicate edges that have *no* coverage.

Render with:
`xdot -f [dot|circo|fdp] Spec_actions.dot`

GraphViz's fdp layout appears to work well to eye-ball-test G for
outliers.  The circo layout is another useful representation.  Lastly,
the actions are clustered into (GraphViz) sub-graphs by their context.
(for TLA+ actions a and b, TLC creates |Proc| contexts if a and b appear
in statements such as \E p \in Proc: a(p) \/ b(p). The GraphViz graph
will have an a, b sub-graph for each context. In other words, A = {a1,
b1, a2, b2, ...} up to |Proc|.

TODO:
* In some cases it is perhaps useful to drop the context dimension
* The system property is a hack
* The weights only counts the chosen successor states out of the set of
generated successors
* Pring the raw adjacency matrix as part of profiling

[Feature][TLC][Simulator][Changelog]