Add operator `TLC!TLCGet("action")`.
```tla
TLC!TLCGet("action") ==
  LET LOCATIONS ==
       [ beginLine: Nat,
         beginColumn: Nat,
         endLine: Nat,
         endColumn: Nat,
         module: STRING ]
  IN CHOOSE act \in
       [ name: STRING,
         location : LOCATIONS ]: TRUE
```

If `TLCGet("action")` is evaluated outside the scope of a TLA+ action of
a behavior spec or if the action is unknown for technical reasons, the
records (name/location) are set to dummy values.

For `TLCGet("action")` to return non-dummy values, TLC internally has to
use its extended state implementation (`tlc2.tool.TLCStateMutExt.java`).
For now, this is the case when TLC runs with "-debugger" or "-simulate".

Note that `TLCGet("action")` remains undocumented in `TLC.tla` until we
have more confidence in its usefulness.

[Feature][TLC]