Move code to create listening socket out of TLCDebugger constructor to let callers acquire its monitor before TLCDebugger spawns its socket. DAP protocol messages from the front-end now block until DebugTool is fully initialized.
Without this change, there exists a race condition where the front-end
connects to TLCDebugger and starts sending protocol messages, such as
"Breakpoints", *before* DebugTool is done parsing the spec (obviously
more likely to occur with larger specs).  This leads to multiple issues
with breakpoints being ignored the most likely one.

In a sane world, TLCDebugger wouldn't start listening before DebugTool
is fully initialized (including parsing the spec).  However, with
current front-end, this opens the door to (front-end) timeouts.  The
real fix is to let TLCDebugger connect to a listening socket spawned by
the front-end, except in the case where a front-end connects to an
already running TLCDebugger.

[Bug][TLC][Debugger]