Show an action's set of successor states under a "Successors" node in the variable view of the debugger.
A user activates an "action breakpoint" by setting
an "in-line breakpoint" on the left-hand side of the action declaration,
i.e., the  SendMsg  in  SendMsg==x'=x  .  Optionally, users may increase
the breakpoint's hit count to only halt evaluation when the number of
successor states is equal to or greater than hit count.  Note that
because of how action breakpoints are set in debugger front-ends,
undeclared actions such as  Spec==Init/\[x'=x]_x  are not supported.


Technical note: For each (declared) action, add a top-most TLA+ debugger
frame to capture the set of successor states "generated" by evaluating
an action. Contrary to all other frames, the breakpoint "fires" on
popping (leaving) the debugger frame (see TLCDebugger.java).

TODO1: Highlight changed variable values in debugger front-ends

TODO2: Add assertions for this feature to the various tests in package
tlc2/debug


[Feature][TLC]