Partial Java module overrides.
Let @Evaluation and @TLAPlusCallable module overrides fall back to pure
TLA+ operator definition of Java code returns null.

This does *not* work for old-school @TLAPlusOperators because the Tool
instance at runtime is not (always) passed to the @TLAPlusOperator
implementation.  However, the main advantage of @TLAPlusOperator is that
Java code doesn't have to manually evaluate arguments, which -for
@TLAPlusOperator- is taken care by OpValue#apply.

[Feature][TLC]