Interactively explore the state-space.
Limitations:
- Supports simulation mode only
- Successors are shown in random order when stepping back
- Limited to the next-state relation (not initial states)
- Visible difference between state and formula stepping granularity
  are the "Step Back" and "Reverse Continue" buttons in the
  debugger's UI
- Stepping over all successors when TLC runs with "-deadlock" will cause
TLC to report a deadlock.

Support hit count (level) for TLCStepActionStackFrames. In other words,
a breakpoint on Next will fire at level/diameter defined as the hit
count in the breakpoint.  When this breakpoint has a hit count, stepping
out (clicking "Step Out" in the front-end) looks like a no-op, because
the breakpoint does not fire for the lower level/diameter.

[Feature][TLC][Debugger]