Refactor LazyValue to help prevent unsafe use of cached values
TLC has had several bugs involving misuse of LazyValue.val.  Instead of
proving every use of that field correct, this change shifts the
management of the field to the LazyValue class itself.  Now callers
must specify the behavior and evaluation mode that they are evaluating
with when reading the cached value.  The cached value is only returned
if it was computed using the same behavior and evaluation mode.  If it
wasn't, it must be recomputed.

This does not GUARANTEE that LazyValue is always used correctly.
Callers can still pass incorrect arguments to LazyValue.getValue() to
cause bugs.  However, it should greatly reduce the likelihood of
further latent LazyValue-related defects.

Pros:

A lot of complexity can be removed without losing any potential
benefits of LazyValue.

Cons:

The LazyValue class now stores several more fields.  It is possible
that this will worsen TLC's performance since many instances of this
class are created during evaluation.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>