Refactor LazyValue to help prevent unsafe use of cached values
MOTIVATION

TLC has had several bugs involving misuse of LazyValue.val.  Instead of
proving every use of that field correct, this change shifts the
management of the field to the LazyValue class itself.  Now callers
must specify the behavior and evaluation mode that they are evaluating
with when reading the cached value.  The cached value is only returned
if it was computed using the same behavior and evaluation mode.  If it
wasn't, it must be recomputed.

This does not GUARANTEE that LazyValue is always used correctly.
Callers can still pass incorrect arguments to LazyValue.getValue() to
cause bugs.  However, it should greatly reduce the likelihood of
further latent LazyValue-related defects.

Pros:

 - A lot of complexity can be removed without losing any potential
   benefits of LazyValue.

Cons:

 - The LazyValue class now stores several more fields.  It is possible
   that this will worsen TLC's performance since many instances of this
   class are created during evaluation.

DETAILS

A key component of this change is the refactored definition of the
condition for WHEN a LazyValue can be re-used in evaluation (see
`LazyValue.getCachedValue`).

First, when can a LazyValue be used with a different TLC state:

A bit of debugging revealed that TLCState.equals() does not behave how
I would expect.  Furthermore, such a check is too strict: TLC often
evaluates a LazyValue in a state where `x=1 /\ y=null` before later
re-evaluating it in a state where `x=1 /\ y=2`.

TLC can use the cached value when it was computed in a superset of the
requested state (see `TLCState.isSubset`).

Second, when can a LazyValue be used with a different EvalControl:

Debugging uncovered that LazyValues are often recomputed using similar
EvalControl settings, such as

    Const | KeepLazy

vs

    Const

We have an opportunity to be more aggressive about caching by
identifying EvalControl settings that are semantically equivalent
(in the sense that they do not affect the meaning of the computed
value).

I have also been careful to handle exceptions due to incomparable
values.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>