Revert the logic changes in commit e7c51944
As we discovered in issue #817, it is not always legal to use an
existing LazyValue in this case.  In particular, it seems to be a
problem when the existing LazyValue's context is not identical to the
requested context `c`.

I have, however, retained the useful new tests and other adjustments
added in that commit.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>