Breakpoint fires for each subexpression when resuming.
* EWD840*Test
The functionality of the 'next'/'continue' command has changed, using
'next' on a line with a breakpoint would act like a 'step-in' command.
This has been corrected, and the 'next command will now perform its
standard operation.
* EchoDebuggerTest
If the debugger is halted on a frame F after breakpoint A fired and one
replaces breakpoint A with new breakpoint B that also fires for the
frame F, B won't fire for the subexpression/frames of F.

This change is similar to the one in
8b7abd12c06700b39b090368ce1b62039dce44b3. However, this changes trades
space for time by lazily checking a frame's parent state
(TLCStackFrame#matches) if the frame matches the breakpoint.  In theory,
this can be expensive if the topmost parent frame has many child frames,
i.e., subexpressions.

Alleviates Github issue #837
https://github.com/tlaplus/tlaplus/issues/837

[Bug][TLC][Debugger]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>