Add TLCExt!TLCGetAndSet operator.
Removes previously commented-out unsuccessful attempts to create a Java
module override for `TLCGetAndSet`.

Let this be a testament to my stupidity to write code when none was
needed. In my defence, I initially assumed that defining `TLCGetAndSet`
was not possible. However, I later realized that it can indeed be
defined using `TLCGet("all")`.

Implements Github issue #858
https://github.com/tlaplus/tlaplus/issues/858

[Feature][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>