Non-determinism makes TLC find a different counterexample.
Introducing the definition `TLCExt!TLCGetAndSet` leads to an increase in
the size of TLC's `util.InternTable` during the initial processing of
this test specification, specifically triggering the
`util.InternTable.grow()` method.  The position of a `UniqueString`
within the table determines the string's tok, which is an input
of the fingerprinting function. Fingerprinting, in turn, affects the
sequence in which liveness checks the different Strongly Connected
Components (SCCs).

[Tests][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>