In `OffHeapDiskFPSet`, invalidate buffered data after concurrent flush
This commit fixes a latent cache inconsistency bug in
`OffHeapDiskFPSet`.  There are no symptoms today, but problems will
arise if we change almost anything in the implementation of
`BufferedRandomAccessFile` or `DiskFPSet`.

Most implementations of `DiskFPSet.Flusher` write directly to the
output file through the provided output file descriptor `outRAF`.  But,
the concurrent flushing algorithm in `OffHeapDiskFPSet` actually writes
to the output file through newly-allocated file descriptors.  As a
result, any buffered data in `outRAF` will be incorrect following a
flush because the on-disk data was modified through different file
descriptors.

Today there are some subtleties in the implementation of
`BufferedRandomAccessFile` that prevent this from causing problems.  It
so happens that when the concurrent flush finishes, there is no
buffered data in `outRAF` to worry about.  However, that state of
affairs is extremely brittle since there is no mechanism to enforce it.

To fix the latent bug, this commit adds a new method that invalidates
the cache in a `BufferedRandomAccessFile`.  The concurrent flush method
calls it at the end of execution to ensure that subsequent reads can
see the data that was written.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>