Flush writes to disk before asserting facts about the disk files
This commit fixes a latent issue in the test suite.

`CommonTestCase.assertNodeAndPtrSizes` makes assertions about the
on-disk state of some files that are internal to `AbstractDiskGraph`.
These checks are useful, but today they only pass because of two
brittle implementation details:

 - `BufferedRandomAccessFile` flushes writes to disk when seeking
   backward in the file.
 - Backward seeks are very common.

The checks will break if we optimize `BufferedRandomAccessFile` to
flush less often or if we change how `AbstractDiskGraph` reads/writes
its internal files.

To make the tests less brittle, this commit introduces a new method
that instructs the liveness checker to flush writes to disk
immediately.  The new method is only called by the tests, and it is
called before making assertions about on-disk state.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>