Add `-lncheck off` to manually disable (potentially expensive) liveness checking.
A temporal property is either a liveness property or a safety property.
For instance, the property <>P is a liveness property, while [](P =>
[]Q) is a safety property (the violation of a safety property is a
finite prefix of a behavior).  TLC checks liveness by searching the
behavior graph for accepting cycles, which amounts to a search for
strongly connected components.  In contrast, safety properties are
verified during the insertion of new nodes into the (partial) behavior
graph.  Note that TLC contains an optimization to check ordinary
invariants (state predicates) during the state-space exploration.

This command-line flag is intentionally left undocumented in hopes that
somebody will eventually find the time to skip liveness checking
automatically if the temporal property is a safety property.

[Feature][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>