Locate the the "StandardModules" folder more reliably
If a given `path` exists on the classpath in multiple places,
`class.getResource(path)` will return an arbitrary one of them.  This
behavior creates a potential problem for `SimpleFilenameToStream` when
running the TLA+ tests:

 - Main classes and test classes are placed in separate folders, and
   both folders are on the classpath.
 - The path "tla2sany/" exists in both locations, since there are some
   test classes in the `tla2sany` package.

As a result, `SimpleFilenameToStream` may decide that the installation
base path is actually "test-classes/tla2sany", which does not contain
a "StandardModules" folder.

This is not a problem in most cases, since in practice `getResource`
returns the _first_ place it found the requested path, and the main
classes go on the classpath first.  But that behavior is not
guaranteed, and furthermore some IDEs like IntelliJ actually put the
test classes first on the classpath.

This patch changes the internals of `SimpleFilenameToStream` to look
for "tla2sany/StandardModules" instead of looking for "tla2sany" and
assuming the arbitrarily-chosen result has a "StandardModules" folder.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>