Document TableauNodePtrTable's bizarre behavior WRT a fingerprints' done state.
Let `fp` be a fingerprint and `t` and `u` be tableau graph nodes.  Assume that
`fp` has been marked as done by calling `TNPT#setDone`.  Further assume that
`<<fp, t>>` and `<<fp, u>>` are added to TNTP in this order.

* Why does adding a GraphNode <<fp, t>> with TNPT#put cause TNPT#isDone(fp) = FALSE?
* Why does marking `<<fp, t>>` as done cause TNPT#isDone(fp) = TRUE, given that
`<<fp, u>>` is still undone.

The first GraphNode of a fingerprint added to TNPT has special meaning. This is
unlikely to be correct, considering that GraphNodes are explored in a
non-deterministic order.

Part of Github issue #971
https://github.com/tlaplus/tlaplus/issues/971

[Tests][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>