Do not override an fp's done state when recording an new <<fp, tidx>> GraphNode (where tidx is some tableau index).
Let s -> t be an action that is inserted into the behavior graph (BG).
Liveness checking marks s as done/explored by adding it to a seen set
S.  It is important to note that S must be distinct from the BFS'
fingerprint set (FPSet) because FPSet is inconsistent with respect to
liveness checking due to the concurrent addition of new states by other
BFS workers (in practice, the elements of both sets are not states but
fingerprints).

After adding s to S, liveness checking processes s->t by creating the
cross product of {t} and the set of consistent nodes in the tableau
graph. The resulting GraphNodes are inserted into BG. If and only if t
is not in S, liveness checking will recursively process all of t's
(state graph) successor states that are likewise not in S.

The implementation had the bug where inserting a GraphNode <<t, ...>>
into the BG would remove t from S as a side effect, violating a
fundamental invariant of the algorithm: [][t \in S => [](t \in S)]_S

This bug was unlikely to occur because state graph nodes are typically
explored in BFS order.

This fix has minimal performance implications, except for the addition
of an extra method parameter to `TableauNodePtrTable#put`. Unless
inlining kicks in, recording a node in the behavior graph will require
one extra stack frame.

Fixes Github issue #971
https://github.com/tlaplus/tlaplus/issues/971

[Bug][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>