Silence noisy `EndOfFileException`
Quitting out of the REPL with Ctrl+D or piping input to the REPL always
results in

    org.jline.reader.EndOfFileException
        at org.jline.reader.impl.LineReaderImpl.readLine(LineReaderImpl.java:697)
        at org.jline.reader.impl.LineReaderImpl.readLine(LineReaderImpl.java:512)
        at tlc2.REPL.runREPL(REPL.java:191)
        at tlc2.REPL.main(REPL.java:236)

This is quite distracting; end-of-file is a normal condition and there
is no reason to see its stacktrace.

This commit silences the noisy output.  It also adds a warning if some
partially-entered input would be ignored because of the end-of-file
(which should be a rare condition).

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>