Remove `Value.assignable()`
This method is not used by the Toolbox, the Community Modules, or TLC
itself.

The `assignable()` method creates problems for Java module overrides
because it is package-private.  A package-private method cannot be
overridden by another class in a different package, meaning that module
overrides cannot create their own `Value` implementations.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>