Added unit test for built-in op init
Before SANY can parse a module, there is a process that initializes the various built-in operators and sets their levels. This is done by calling `Configuration.load()` to set the operator name & arity then `BuiltInLevel.load()` to set the level constraints. This work adds a unit tests covering the operation of these two functions both to check that their action is as expected, and as protection to ensure any future refactoring of these functions preserves their action.

This is all test code with the exception of a visibility change on an internal class (which was probably meant to be publicly-visible, because instances of it are returned from a publicly-visible method) and addition of two methods exposing some internal level constraint information for testing purposes.

Issue #1008 was found during this work.

Signed-off-by: Andrew Helwer <2n8rn1w1f@mozmail.com>