Support running the TLA+ debugger with multiple workers.
Previously, the debugger could only run with a single worker (`-workers
1`), limiting its usefulness for large models. This commit introduces
support for running the debugger with multiple workers (`-workers N`).

The debugger attaches to one worker, while the other workers continue
without being directly debugged. This setup reduces debugging overhead,
such as the creation of stack frames, on the non-debugged workers.

This is a provisional feature for now that has to be explicitly activated

[Feature][TLC][Changelog][Debugger]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>