Correct the predecessors and level of *all* states after dropping *finite* stuttering steps.
To simplify the counterexample, simulation removes finite stuttering
steps (1f9f3cb59854937547a24d59e057cec02d476d5f), reducing traces
like <<a, a, a, b, b, c>> to <<a, b, c>>. This PR ensures that e.g.
TLCGet("level") correctly evaluates in an ALIAS or POSTCONDITION to
1, 2, and 3 for the three respective states. Previously, it
incorrectly evaluated to something like 1, 4, and 6. This is done by
refactoring the code to consistently append the final TLCState to
the stateTrace (see previous commit), and iterating over the trace
and resetting the predecessor for each state in the sequence (except
the init state).

[Bug][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>