NullPointerException when last state of the trace is *not* in the model.
If the final state of a counterexample is excluded from the model
because state- or action-constraints evaluate to false, it is not
written to the trace file.  This omission causes a NullPointerException
when attempting to read the final state from the trace file during
counterexample reconstruction.  To address this issue, the final state
is now explicitly appended to the counterexample to ensure completeness
and avoid runtime errors.

[Bug][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>