Fix Typo in README
In this PR, I have corrected a typo in the project's README file, changing "t0" to "to". Although the change is minor, I hope it helps improve the accuracy and readability of the documentation. Thank you to the maintainers for their hard work!
---
Feel free to modify it further if needed.