ArrayIndexOutOfBoundsException in TLA+ debugger during simulation if simulator eliminated finite stuttering from current trace.
```
Caused by: java.lang.ArrayIndexOutOfBoundsException: Index 82 out of
bounds for length 82
        at tlc2.tool.StateVec.elementAt(StateVec.java:78)
        at tlc2.tool.SimulationWorker.getTraceInfo(SimulationWorker.java:811)
        at tlc2.tool.Simulator.getTraceInfo(Simulator.java:816)
        at tlc2.debug.TLCStateStackFrame.lambda$getVariables$2(TLCStateStackFrame.java:163)
        at tlc2.tool.impl.DebugTool.eval(DebugTool.java:686)
        at tlc2.debug.TLCStateStackFrame.getVariables(TLCStateStackFrame.java:141)
        at tlc2.debug.TLCDebugger.variables(TLCDebugger.java:547)
        at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:103)
        ... 13 more
```

Related to git commit 1f9f3cb59854937547a24d59e057cec02d476d5f. Due to
this commit, the "Action" node in the debugger's variable view is no
longer guaranteed to show ordinal n+1 if the topmost state in the
debugger's "Trace" node shows n. Instead, depending on the number of
stuttering steps, that number can be lower. This is a feature. :-)

[Bug][TLC][Debugger]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>