Make deserializing previously unseen strings work
Currently, if TLC tries to load a string value from disk that its current instance has not already "seen", it will crash.
This is because loading strings from disk does not update the string intern table. The current code only reads from it, and that leads to the impossible situation of TLC needing to know a string already before it can load it.

This behavior was not originally a problem, because the binary serialization mechanism was written for the state queue and checkpointing. Those correctly load a separately-serialized string table that has all the right entries, making loading an "unknown" string impossible.

When this feature was extended to support loading arbitrary values in the community modules, however, "unknown" strings become not only possible, but quite likely in that scenario.

This fix is to add a `readExternal` method that replaces the current `read(Map)` that can only make lookups into a read-only string table. Instead, `readExternal` adds strings to the table as needed. The effective change is 1 LoC.

To avoid breaking older community modules versions (like the one in test dependencies right now), the `read(Map)` method is left in as a deprecated proxy to `readExternal`.

Signed-off-by: fhackett <fhackett.py@gmail.com>