Refactor `SimpleFilenameToStream` for correctness and clarity
Fixes #1114.

This change has multiple purposes.  The primary purpose is to allow
`SimpleFilenameToStream` to resolve the standard library files in
other environments, such as when the tools are compiled as a Graal
native image or as a jlink'd executable.

Previously `SimpleFilenameToStream` relied on resolving the resource
folder `tla2sany/StandardModules` to a URL and then examining the URL's
scheme, which is not defined behavior:

 - In Graal the scheme is `resource:` and not `jar:`
 - In jlink the directory always resolves to `null`, even if entries
   within it are accessible.

To fix the problem, the search logic in `SimpleFilenameToStream` has
been broken out into different "ResourceLocator" classes, each of which
does very little, but which can be composed into a sequence of
locations to search.  (I had originally thought to make a set of
composable implementations of `FilenameToStream`, but the contract for
that interface has too many warts.  It was much easier to replace the
internals of `SimpleFilenameToStream` than to rework the contracts of
existing methods, especially since the Toolbox uses them.)

The fix is in `ClasspathResourceLocator.java`, where resources are
resolved directly using `classLoader.getResource(...)` and there is no
"pre-resolve" step on the prefix directory.

As a secondary purpose, I believe these changes also improve the
clarity of TLC's module resolution process---or at least, they
eliminate a lot of bizarre behaviors:

 - Clearer typing (everything is now URLs instead of Strings-that-are-
   either-URLs-or-paths).
 - Fewer resource lookups (the old code ran `getResource` _and_
   `getResourceAsStream` after examining the prefix string).
 - Reduced redundant calls to `File.exists()` (the `read()` method will
   always return an existing file).
 - Straightforward for-each loops (no special-casing on index 0, no
   special case on libaryPaths.length-1, no reliance on the last lookup
   path equaling the standard module folder, etc).

We have good test coverage of `SimpleFilenameToStream`, and the
existing tests pass, along with some new ones that pass both before and
after this change.

There is a bonus fix here as well: `isStandardModule` now works even
when TLC is running out of a jar, and its implementation has been
simplified to reduce duplication with `TLAFile.isLibraryModule()`.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>