Warning: DiskFPSet.mergeNewEntries: xxx is already on disk with hangup following.
A data corruption (either in memory or on disk) or a bug in the
fingerprint set implementation caused a fingerprints stored on disk to
match one in memory. Despite this issue, model checking can continue
without any compromise to the soundness or completeness of the safety
checks. At worst, TLC ends up checking the same state multiple times due
to the duplicate fingerprints.

Part of Github issue #1112
https://github.com/tlaplus/tlaplus/issues/1112

[Bug][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>