Change to instance type with higher storage capacity when running performance benchmarks
The storage capacity of the previous instance type was insufficient, resulting in the benchmarks failing with the error message: `No space left on device`. To address this, the `c6gd.4xlarge` instance type [1] is being used. This instance type includes local/ephemeral storage, which is conveniently mounted at `GITHUB_WORKSPACE` [2].

[1] https://aws.amazon.com/ec2/instance-types/c6g/
[2] https://github.com/runs-on/runs-on/issues/144#issue-2593981972

[Build]