Clean up (and inline) `SimpleFilenameToStream.getLibraryPaths()`
Follows up on #1116.

Ref: https://github.com/tlaplus/tlaplus/pull/1116#discussion_r1911373993
Ref: https://github.com/tlaplus/tlaplus/pull/1116#discussion_r1911809892

The `getLibraryPaths()` method is inefficient, poorly formatted, and
needlessly confusing: with explicitly-provided library paths it will
construct a colon-separated string in memory, only to immediately split
that string back into an array.

This commit simplifies the method to the point where it makes sense to
just inline it.  The `SimpleFilenameToStream` constructor now makes it
a bit more obvious what is being searched.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>