Fix path resolution on Windows
Fixes a bug in commit eaeb041d.  On Windows, path URLs have a form like

    file:/C:/path

Taking `URL.getPath()` gives back

    /C:/path

which is not a legal Windows path.  However, there is built-in
functionality for converting URLs to paths, which I should have used
from the beginning. The built-in functionality correctly recovers the
path

    C:\path

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>