LazyValue no longer inherits from Value
LazyValue supports none of the functionality of Value:
 - it can't be compared for equality
 - it can't be tested for membership
 - it can't be serialized
 - it can't be converted to other Value types
 - it can't be passed to Java operator overrides

It seems tempting to treat LazyValue as a not-yet-evaluated value, but
that doesn't work out in practice.  The meaning of a LazyValue depends
on the current behavior and whether it is being used under a prime
operator or inside `ENABLED`.  TLC does not (and has never) allowed
LazyValue inside another LazyValue or inside other composite values
like SetEnumValue.

This change should make us a little more confident in TLC's existing
(and future) code: there is greatly reduced risk of LazyValue "leaking"
into places it doesn't belong.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>