Add a warning if a state-level formula is used as a PROPERTY.
"The formula %1% is a %2% formula, but it was used as a PROPERTY (or
PROPERTIES), where a temporal formula is typically expected.
Non-temporal formulas used as PROPERTY or PROPERTIES are only checked in
the initial state. To verify that the %2% formula %1% holds in all
states of every behavior, use INVARIANT %1% instead. Alternatively,
applying the \"always\" temporal operator (□) to the "%2% formula %1%
changes it into a temporal formula, asserting that %1% holds in all
states of every behavior."

[Feature][TLC]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>