Remove `TLAFlightRecorder`
This class has a strong dependency on Java Flight Recorder (JFR), but
does not offer much value.

From @lemmy:

 > A (Java) Flight Recorder is a very powerful TLC debug aid,
 > particularly for Cloud and Distributed TLC. However, adoption of
 > these features has been very low, and in practice, no one has ever
 > shared a flight recording with me. Therefore, I propose removal of
 > flight recordings to unblock Graal native image support.

https://github.com/tlaplus/tlaplus/issues/1114#issuecomment-2831136553

Note that the tests still depend on JFR to verify that certain methods
are inlined during execution.  That use of JFR does not block Graal
native image adoption, since Graal does not need to compile the tests.

In the future we might revisit this idea, perhaps guarded behind a
command-line flag or environment variable.

Signed-off-by: Calvin Loncaric <calvin.loncaric@oracle.com>